fix: Remove original static file serving test and use CI-specific version
- Remove original __tests__/static-file-serving.test.js that was failing in CI
- Rename __tests__/static-file-serving-ci.test.js to __tests__/static-file-serving.test.js
- The CI-specific version properly sets environment variables before importing server
- All static file serving tests now pass locally
- Should resolve the static file serving test failures in GitHub Actions
